+++ neubot/mod_bittorrent/__init__.py
@@ -1,4 +1,4 @@
-# neubot/bittorrent/__init__.py
+# neubot.mod_bittorrent/__init__.py
 
 #
 # Copyright (c) 2011 Simone Basso <bassosimone@gmail.com>,
@@ -34,12 +34,12 @@ import sys
 if __name__ == "__main__":
     sys.path.insert(0, ".")
 
-from neubot.bittorrent.peer import PeerNeubot
-from neubot.bittorrent.server import ServerPeer
+from neubot.mod_bittorrent.peer import PeerNeubot
+from neubot.mod_bittorrent.server import ServerPeer
 from neubot.lib_http.server import HTTP_SERVER
 from neubot.lib_net.poller import POLLER
 
-from neubot.bittorrent import config
+from neubot.mod_bittorrent import config
 from neubot.config import CONFIG
 
 from neubot import log
@@ -88,9 +88,9 @@ def main(args):
     try:
         options, arguments = getopt.getopt(args[1:], '6A:fp:v')
     except getopt.error:
-        sys.exit('usage: neubot bittorrent [-6fv] [-A address] [-p port]')
+        sys.exit('usage: neubot.mod_bittorrent [-6fv] [-A address] [-p port]')
     if arguments:
-        sys.exit('usage: neubot bittorrent [-6fv] [-A address] [-p port]')
+        sys.exit('usage: neubot.mod_bittorrent [-6fv] [-A address] [-p port]')
 
     prefer_ipv6 = 0
     address = 'master.neubot.org'
